# docs/Mark-Layer2-API-Contracts.yaml
openapi: 3.0.3
info:
  title: MARK Layer-2 and Wellbeing API Contracts
  version: 1.0.0
  description: >-
    Contract-first API specification for MARK workplace assistant tools and
    wellbeing reminder operations.
servers:
  - url: /api/v1
    description: Base path
...
paths:
...
  /wellbeing/reminders:
    get:
      tags: [Wellbeing]
      summary: List active and paused reminders
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Reminder list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListRemindersResponse'
    post:
      tags: [Wellbeing]
      summary: Create new reminder
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateReminderRequest'
      responses:
        '201':
          description: Reminder created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReminderResponse'
  /wellbeing/reminders/{reminderId}:
    patch:
      tags: [Wellbeing]
      summary: Update reminder state or schedule
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: reminderId
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateReminderRequest'
      responses:
        '200':
          description: Reminder updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReminderResponse'
    delete:
      tags: [Wellbeing]
      summary: Cancel reminder
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: reminderId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Reminder cancelled
  /wellbeing/check-ins/daily:
    post:
      tags: [Wellbeing]
      summary: Record daily check-in response
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DailyCheckinRequest'
      responses:
        '201':
          description: Check-in recorded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DailyCheckinResponse'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    CreateReminderRequest:
      type: object
      required: [reminderType, title, message, scheduleKind]
      properties:
        reminderType:
          type: string
          enum: [break, medicine, meeting, custom]
        title:
          type: string
        message:
          type: string
        scheduleKind:
          type: string
          enum: [one_time, daily, weekly, cron]
        runAt:
          type: string
          format: date-time
        cronExpr:
          type: string
        timezone:
          type: string
          default: UTC
    UpdateReminderRequest:
      type: object
      properties:
        title:
          type: string
        message:
          type: string
        status:
          type: string
          enum: [active, paused, cancelled]
        runAt:
          type: string
          format: date-time
        cronExpr:
          type: string
    ReminderResponse:
      type: object
      properties:
        reminderId:
          type: string
          format: uuid
        reminderType:
          type: string
        scheduleKind:
          type: string
        status:
          type: string
        timezone:
          type: string
        nextTriggerAt:
          type: string
          format: date-time
    ListRemindersResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ReminderResponse'
    DailyCheckinRequest:
      type: object
      required: [mood, message]
      properties:
        mood:
          type: string
          enum: [great, good, okay, low, stressed]
        message:
          type: string
        wantsFollowup:
          type: boolean
    DailyCheckinResponse:
      type: object
      properties:
        checkinId:
          type: string
          format: uuid
        sentimentLabel:
          type: string
          enum: [positive, neutral, negative]
        suggestedNextStep:
          type: string